Add an initial implementation of the "pretty indent" feature (issue #20)
Lot's of things are still missing (right now, only the SELECT statement is covered), but the
basic infrastructure is there and seems to work reasonably well.

Obviously the "pretty is in the eye of the beholder", and the formatting style is currently
very biased on my own taste, but the idea is that it could be driven by a set of options, once
the dust settles.